fix: 게임 재시작 여부 검증 담당 클래스를 수정
Application 단위에서 R, Q 판단 후 종료, 재시작을 위하여 BridgeGame이 아닌 InputView에서 검증을 담당하게 수정.
